requirements.txt
fastapi
uvicorn
supabase
python-dotenv
httpx